Update and fix bad exception handling
This is the first commit in a series that will bring this library up to
Python 3 compatibility. Catching all exceptions and either swallowing
them or changing the exception type really isn't going to be helpful for
this (or in general) because it removes the stack trace and exception
type. If it is really believed that these exceptions are good, I would
suggest adding them back in when this process is complete. Some except
blocks were modified to change the syntax as the current syntax is
removed in Python 3 (and was deprecated in like 2.4 or something in the
80s).