Fix cases of division producing floats in Python3
Python 3's division operator produces floats. If this is not desired,
there is a // operator that will produce the floor of this number, like
programmers traditionally expect. I identified cases where this may
cause a problem by running the tests with the -3 argument. At this point
I discovered that the test coverage for this project isn't that good.
But I examined other cases manually and think this probably fixes most
of the issues. *shrug*

from __future__ import division was added to each file to prevent
regression.