Switch files to using unicode literals
Python 3 is all unicode, all the time, so a Python 3 compatible version
of this library must handle unicode correctly. This commit was produced
by going file by file, adding the unicode literals import, and then
fixing each failed test on a case by base basis. There are still
complications with Python 3 unicode however, because it seems numpy
expects unicode strings for dtype definitions when using Python 3, and
ascii when using Python 2. In fact, I'm sure there are still many
problems not covered by the tests, but hopefully this is a good start.