Bring project up to Python 3 compatibility
All the tests pass so it must be 100% functioning, right???

Many changes were necessary to get the tests to pass. In particular, all
the parts where data is read and written needed to be changed quite a
bit. Maps and zips were replaced with list comprehensions (Map and zip
return a generator in Python 3, which may cause problems if lists are
expected). Additionally, other style issues were worked on. __del__
functions that close file handles were deleted, as it was causing
unhelpful test failure messages, and is not a good way to close files as
there is no control over when it would run, or even if it will run at
all.