Clean up style of return and raise statements
Return is not a function and it's unlikely that it ever will be. It is
flow control. Unnecessary parentheses clutter the syntax, making it less
clear. Empty return statements at the end of the function do not alter
the meaning of the program or communicate anything so they are removed
in this commit.